Fix settings during search steps to materialized
