+++ Api/AdministratorServices/AgentSystemSettingsManagementService.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using Api.Services.Internal;
 using CSharpFunctionalExtensions;
 using HappyTravel.Edo.Api.Extensions;
 using HappyTravel.Edo.Api.Infrastructure.FunctionalExtensions;
@@ -6,7 +7,6 @@ using HappyTravel.Edo.Api.Models.Management.AuditEvents;
 using HappyTravel.Edo.Api.Models.Settings;
 using HappyTravel.Edo.Api.Services.Management;
 using HappyTravel.Edo.Common.Enums;
-using HappyTravel.Edo.Common.Enums.AgencySettings;
 using HappyTravel.Edo.Data;
 using HappyTravel.Edo.Data.Agents;
 using Microsoft.EntityFrameworkCore;
@@ -16,11 +16,14 @@ namespace HappyTravel.Edo.Api.AdministratorServices
     public class AgentSystemSettingsManagementService : IAgentSystemSettingsManagementService
     {
         public AgentSystemSettingsManagementService(EdoContext context,
-            IManagementAuditService managementAuditService, IAgencySystemSettingsManagementService agencyManagementService)
+            IManagementAuditService managementAuditService,
+            IAgencySystemSettingsManagementService agencyManagementService,
+            IInternalSystemSettingsService internalSystemSettingsService)
         {
             _context = context;
             _managementAuditService = managementAuditService;
             _agencyManagementService = agencyManagementService;
+            _internalSystemSettingsService = internalSystemSettingsService;
         }
 
 
@@ -62,8 +65,7 @@ namespace HappyTravel.Edo.Api.AdministratorServices
 
 
         public Task<Result<AgentAccommodationBookingSettings>> GetAvailabilitySearchSettings(int agentId, int agencyId)
-            => CheckRelationExists(agentId, agencyId)
-                .Map(() => GetSettings(agentId, agencyId));
+            => _internalSystemSettingsService.GetAgentMaterializedSearchSettings(agentId, agencyId);
 
 
         public async Task<Result> DeleteAvailabilitySearchSettings(int agentId, int agencyId)
@@ -91,71 +93,17 @@ namespace HappyTravel.Edo.Api.AdministratorServices
         }
 
 
-        private async Task<Result> CheckRelationExists(int agentId, int agencyId)
-            => await _context.AgentAgencyRelations.AnyAsync(r => r.AgentId == agentId && r.AgencyId == agencyId)
-                ? Result.Success()
-                : Result.Failure("Could not find specified agent in given agency");
+        private Task<Result> CheckRelationExists(int agentId, int agencyId)
+            => _internalSystemSettingsService.CheckRelationExists(agentId, agencyId);
 
 
         private Task<AgentSystemSettings> GetOrDefault(int agentId, int agencyId)
             => _context.AgentSystemSettings.SingleOrDefaultAsync(s => s.AgentId == agentId && s.AgencyId == agencyId);
 
 
-        private async Task<AgentAccommodationBookingSettings> GetSettings(int agentId, int agencyId)
-        {
-            var (_, isAgencySettingsFailure, materializedAgencySettings) = await _agencyManagementService.GetAvailabilitySearchSettings(agencyId);
-            var agentSettings = (await _context.AgentSystemSettings.SingleOrDefaultAsync(s => s.AgentId == agentId && s.AgencyId == agencyId))
-                ?.AccommodationBookingSettings;
-
-            if (agentSettings is null)
-            {
-                if (isAgencySettingsFailure)
-                    return new AgentAccommodationBookingSettings
-                    {
-                        AdditionalSearchFilters = new(),
-                        AprMode = AprMode.Hide,
-                        CustomDeadlineShift = 0,
-                        IsDirectContractFlagVisible = false,
-                        IsSupplierVisible = false,
-                        PassedDeadlineOffersMode = PassedDeadlineOffersMode.Hide
-                    };
-
-                return new AgentAccommodationBookingSettings
-                {
-                    AdditionalSearchFilters = new(),
-                    AprMode = materializedAgencySettings.AprMode,
-                    CustomDeadlineShift = materializedAgencySettings.CustomDeadlineShift,
-                    IsDirectContractFlagVisible = materializedAgencySettings.IsDirectContractFlagVisible,
-                    IsSupplierVisible = materializedAgencySettings.IsSupplierVisible,
-                    PassedDeadlineOffersMode = materializedAgencySettings.PassedDeadlineOffersMode
-                };
-            }
-
-            var aprMode = agentSettings.AprMode > materializedAgencySettings.AprMode
-                ? materializedAgencySettings.AprMode
-                : agentSettings.AprMode;
-
-            var passedDeadlineOffersMode = agentSettings.PassedDeadlineOffersMode > materializedAgencySettings.PassedDeadlineOffersMode
-                ? materializedAgencySettings.PassedDeadlineOffersMode
-                : agentSettings.PassedDeadlineOffersMode;
-
-            var isDirectContractFlagVisible = agentSettings.IsDirectContractFlagVisible && materializedAgencySettings.IsDirectContractFlagVisible;
-            var isSupplierVisible = agentSettings.IsSupplierVisible && materializedAgencySettings.IsSupplierVisible;
-
-            return new AgentAccommodationBookingSettings
-            {
-                AdditionalSearchFilters = agentSettings.AdditionalSearchFilters,
-                AprMode = aprMode,
-                CustomDeadlineShift = agentSettings.CustomDeadlineShift,
-                IsDirectContractFlagVisible = isDirectContractFlagVisible,
-                IsSupplierVisible = isSupplierVisible,
-                PassedDeadlineOffersMode = passedDeadlineOffersMode
-            };
-        }
-
-
         private readonly EdoContext _context;
         private readonly IManagementAuditService _managementAuditService;
         private readonly IAgencySystemSettingsManagementService _agencyManagementService;
+        private readonly IInternalSystemSettingsService _internalSystemSettingsService;
     }
 }
