+++ Api/Services/Agents/AgencySystemSettingsService.cs
@@ -1,31 +1,32 @@
 using System.Threading.Tasks;
+using Api.Services.Internal;
 using CSharpFunctionalExtensions;
-using HappyTravel.Edo.Api.Extensions;
-using HappyTravel.Edo.Api.Models.Agents;
-using HappyTravel.Edo.Common.Enums.AgencySettings;
 using HappyTravel.Edo.Data;
 using HappyTravel.Edo.Data.Agents;
-using Microsoft.EntityFrameworkCore;
 
 namespace HappyTravel.Edo.Api.Services.Agents
 {
     public class AgencySystemSettingsService : IAgencySystemSettingsService
     {
-        public AgencySystemSettingsService(EdoContext context)
+        public AgencySystemSettingsService(EdoContext context,
+            IInternalSystemSettingsService internalSystemSettingsService)
         {
             _context = context;
+            _internalSystemSettingsService = internalSystemSettingsService;
         }
 
 
         public async Task<Maybe<AgencyAccommodationBookingSettings>> GetAccommodationBookingSettings(int agencyId)
         {
-            var settings = await _context.AgencySystemSettings
-                .SingleOrDefaultAsync(s => s.AgencyId == agencyId);
+            var (_, isFailure, settings) = await _internalSystemSettingsService.GetAgencyMaterializedSearchSettings(agencyId);
+            if (isFailure)
+                return Maybe<AgencyAccommodationBookingSettings>.None;
 
-            return settings?.AccommodationBookingSettings;
+            return settings;
         }
 
 
         private readonly EdoContext _context;
+        private readonly IInternalSystemSettingsService _internalSystemSettingsService;
     }
 }
